% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron.R
\name{chron_as_sf}
\alias{chron_as_sf}
\title{Convert a table of chronological data to an sf object}
\usage{
chron_as_sf(x, crs = sf::st_crs(4626))
}
\arguments{
\item{x}{data.frame retrieved with \code{\link[=chron_data]{chron_data()}}.}

\item{crs}{Desired coordinate reference system in a format understood by
\code{\link[sf:st_crs]{sf::st_crs()}}. Defaults to latitude/longitude on the WGS84 ellipsoid
(EPSG:4326).}
}
\value{
An \code{sf} object.
}
\description{
Adds a simple features geometry column (see \link[sf:sf]{sf::st_sf}) to chronological
data from XRONOS, using the latitude and longitude columns. Rows with missing
or invalid coordinates will be dropped with a warning.
}
\examples{
x <- chron_data(country = "Switzerland")
chron_as_sf(x)
}
