% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron.R
\name{chron_data}
\alias{chron_data}
\title{Get chronological data from XRONOS}
\usage{
chron_data(..., .everything = NA)
}
\arguments{
\item{...}{(Optional) One or more named arguments specifying filter variables
and values to include. Supported filters are \code{"labnr"}, \code{"site"},
\code{"site_type"}, \code{"country"}, \code{"feature"}, \code{"material"}, and \code{"species"}.}

\item{.everything}{If \code{TRUE}, suppresses interactive mode prompt when
retrieving all records from XRONOS (see details).}
}
\description{
Retrieves chronological data from XRONOS. \code{chron_data()} retrieves all
records. Additional parameters can be used to filter the results.
}
\details{
To reduce unnecessary server load, \code{get_xronos()} (without any filters) will
prompt for confirmation in interactive mode interactive mode. Set
\code{.everything = TRUE} to suppress this.

\code{country} should be a valid
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-21}{ISO two-letter country code}.
Otherwise, the function will try to interpret as such using
\code{\link[countrycode:countryname]{countrycode::countryname()}}, or return an error if any values of \code{country}
cannot be matched.
}
\examples{
# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))
}
