% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xronos-package.R
\docType{package}
\name{xronos-package}
\alias{xronos}
\alias{xronos-package}
\title{xronos: Client for the 'XRONOS' Chronological Database}
\description{
A programmatic interface to 'XRONOS' \url{https://xronos.ch}, an open data infrastructure for the backbone of the archaeological chronology, including radiocarbon and dendrochronological dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.xronos.ch}
  \item \url{https://github.com/xronos-ch/xronos.R}
  \item Report bugs at \url{https://github.com/xronos-ch/xronos.R/issues}
}

}
\author{
\strong{Maintainer}: Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID})

Authors:
\itemize{
  \item Martin Hinz \email{martin.hinz@iaw.unibe.ch} (\href{https://orcid.org/0000-0002-9904-6548}{ORCID})
}

Other contributors:
\itemize{
  \item Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [contributor]
  \item Swiss National Science Foundation [funder]
}

}
\keyword{internal}
