% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xronos-api.R
\name{xronos_api_url}
\alias{xronos_api_url}
\alias{xronos_query}
\title{Query the XRONOS API}
\usage{
xronos_api_url(version = "v1")

xronos_query(query, api_url = xronos_api_url())
}
\arguments{
\item{version}{Version of the API to use. Currently has no effect.}

\item{query}{Valid query string.}

\item{api_url}{Root address of the XRONOS API. \code{xronos_api_url()} provides
the default value.}
}
\value{
Parsed JSON response.
}
\description{
Directly query the XRONOS API. See \url{https://xronos.ch/api} for supported
formats.
}
\examples{
xronos_query("query_labnr=AAR-1847")
}
