% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xronos-api.R
\name{xronos_url}
\alias{xronos_url}
\alias{xronos_api_url}
\title{XRONOS web address}
\usage{
xronos_url()

xronos_api_url(version = "v1", base_url = xronos_url())
}
\arguments{
\item{version}{Version of the API to use. Currently has no effect.}

\item{base_url}{Base URL of XRONOS.}
}
\description{
Returns the current base URL of XRONOS, or the address of the API version
specified.
}
\examples{
xronos_url()

xronos_api_url()
}
