% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron.R
\name{chron_data}
\alias{chron_data}
\title{Get chronological data from XRONOS}
\usage{
chron_data(..., .everything = NA)
}
\arguments{
\item{...}{(Optional) One or more named arguments specifying filter variables
and values to include. See \code{\link[=xronos_query]{xronos_query()}} for a list of filters supported
by the XRONOS API.}

\item{.everything}{If \code{TRUE}, suppresses interactive mode prompt when
retrieving all records from XRONOS (see details).}
}
\value{

}
\description{
Retrieves chronological data from XRONOS. \code{chron_data()} retrieves all
records. Additional parameters can be used to filter the results.
}
\details{
To reduce unnecessary server load, \code{get_xronos()} (without any filters) will
prompt for confirmation in interactive mode interactive mode. Set
\code{.everything = TRUE} to suppress this.
}
\examples{
# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))
}
