% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xronos-api.R
\name{xronos_query}
\alias{xronos_query}
\title{Query the XRONOS API}
\usage{
xronos_query(filter, values)
}
\arguments{
\item{filter}{Name of a variable or vector of variables to filter by.
Supported filters are \code{"labnr"}, \code{"site"}, \code{"site_type"}, \code{"country"},
\code{"feature"}, \code{"material"}, and \code{"species"}.}

\item{values}{A vector of values to include. If more than one \code{filter} is
used, must be a list of the same length.}
}
\value{
Parsed JSON response.
}
\description{
Constructs a query request to the XRONOS API by combining filter variables
and values.
}
\examples{
xronos_query(c("site", "material"), list("Stonehenge", c("bone", "charcoal")))
}
